requirements-test.txt
pytest
pytest-xdist
pytest-split
requests
fastapi
uvicorn[standard]
pydantic
pyautogen==0.2.2
